refactor(quota): modularize quota fetchers to providers package and restrict to oauth
