+++ packages/providers/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./oauth.js";
+export * from "./oauth/index.js";
+export * from "./quota/index.js";
 export * from "./registry.js";
 export * from "./circuitBreaker.js";
